+++ src/main/java/com/bedsbutendgame/config/BbeConfig.java
@@ -1,6 +1,9 @@
 package com.bedsbutendgame.config;
 
 public final class BbeConfig {
-	public boolean disablePhantoms = true;
-	public boolean nightmares = true;
+	public static final boolean DEFAULT_DISABLE_PHANTOMS = true;
+	public static final int DEFAULT_NIGHTMARE_CHANCE = 35;
+
+	public boolean disablePhantoms = DEFAULT_DISABLE_PHANTOMS;
+	public int nightmareChance = DEFAULT_NIGHTMARE_CHANCE;
 }